General

Lux.apply — Function

apply(model::AbstractExplicitLayer, x, ps::Union{ComponentArray,NamedTuple}, st::NamedTuple)

Simply calls model(x, ps, st)

Lux.setup — Function

setup(rng::AbstractRNG, l::AbstractExplicitLayer)

Shorthand for getting the parameters and states of the layer l. Is equivalent to (initialparameters(rng, l), initialstates(rng, l)).

Parameters

Lux.initialparameters — Function

initialparameters(rng::AbstractRNG, l)

Generate the initial parameters of the layer l.

Lux.parameterlength — Function

parameterlength(l)

Return the total number of parameters of the layer l.

States

Lux.initialstates — Function

initialstates(rng::AbstractRNG, l)

Generate the initial states of the layer l.

Lux.statelength — Function

statelength(l)

Return the total number of states of the layer l.

Lux.testmode — Function

testmode(st::NamedTuple, mode::Bool=true)

Make all occurances of training in state st !mode

Lux.trainmode — Function

trainmode(x::Any, mode::Bool=true)

Make all occurances of training in state st mode

Lux.update_state — Function

update_state(st::NamedTuple, key::Symbol, value; layer_check=_default_layer_check(key))

Recursively update all occurances of the key in the state st with the value.

Index

Lux.apply

Lux.initialparameters

Lux.initialstates

Lux.parameterlength

Lux.setup

Lux.statelength

Lux.testmode

Lux.trainmode

Lux.update_state