Experimental Features

Warning

These features are relatively new additions to Lux, and as such haven't been pressure tested enough. We recommend users to carefully go through these docs, since they highlight some of the common gotchas when using these features. As always, if something doesn't work and the documentation doesn't explicitly mention that it shouldn't work, please file an issue

Lux.@layerdef API

Lux.@layerdef — Macro

An Experimental API to make it easier to write Lux models.

Example

using Lux, Random, Zygote, NNlib

Lux.@layerdef function MyCustomModel((x, y)::T) where {T}
    @assert size(x, 1) == 2 && size(y, 1) == 2
    a = Dense(3, 3, relu)(Dense(2, 3)(x) .+ Dense(2, 3, relu)(x)) .- Dense(2, 3, gelu)(x)
    b = Dense(3, 2, sigmoid)(a)
    c = Dense(2, 2, relu)(y)
    return b .+ c
end

model = MyCustomModel()

ps, st = Lux.setup(Random.default_rng(), model)
x, y = randn(Float32, 2, 1), randn(Float32, 2, 1)

model((x, y), ps, st)

Verifying Generated Code

Tip

Since this is an highly experimental feature, we recommend always manually verifying the generated model. This can be done simply using @macroexpand Lux.@layerdef ...

julia> @macroexpand Lux.@layerdef function MyCustomModel((x, y)::T) where {T}
           @assert size(x, 1) == 2 && size(y, 1) == 2
           a = Dense(3, 3, relu)(Dense(2, 3)(x) .+ Dense(2, 3, relu)(x)) .- Dense(2, 3, gelu)(x)
           b = Dense(3, 2, sigmoid)(a)
           c = Dense(2, 2, relu)(y)
           return b .+ c
       end
quote
    struct MyCustomModel{__dense_1_type, __dense_2_type, __dense_3_type, __dense_4_type, __dense_5_type, __dense_6_type} <: Lux.AbstractExplicitContainerLayer{(:dense_1, :dense_2, :dense_3, :dense_4, :dense_5, :dense_6)}
        dense_1::__dense_1_type
        dense_2::__dense_2_type
        dense_3::__dense_3_type
        dense_4::__dense_4_type
        dense_5::__dense_5_type
        dense_6::__dense_6_type
    end
    function MyCustomModel(; )
        return MyCustomModel(Dense(3, 3, relu), Dense(2, 3), Dense(2, 3, relu), Dense(2, 3, gelu), Dense(3, 2, sigmoid), Dense(2, 2, relu))
    end
    function (mandrill::MyCustomModel)((x, y)::T, ps::Union{Lux.ComponentArray, NamedTuple}, st::NamedTuple; ) where T
        if size(x, 1) == 2 && size(y, 1) == 2
            nothing
        else
            Base.throw(Base.AssertionError("size(x, 1) == 2 && size(y, 1) == 2"))
        end
        (wildebeest, camel) = mandrill.dense_4(x, ps.dense_4, st.dense_4)
        (antelope, gaur) = mandrill.dense_3(x, ps.dense_3, st.dense_3)
        (rhinoceros, ibis) = mandrill.dense_2(x, ps.dense_2, st.dense_2)
        (mallard, jellyfish) = mandrill.dense_1(rhinoceros .+ antelope, ps.dense_1, st.dense_1)
        a = mallard .- wildebeest
        (hamster, monkey) = mandrill.dense_5(a, ps.dense_5, st.dense_5)
        b = hamster
        (wasp, sanddollar) = mandrill.dense_6(y, ps.dense_6, st.dense_6)
        c = wasp
        return (b .+ c, NamedTuple{(:dense_1, :dense_2, :dense_3, :dense_4, :dense_5, :dense_6)}((jellyfish, ibis, gaur, camel, monkey, sanddollar)))
    end
end

What not to do while writing layerdefed layers?

Here are some common cases where @layerdef will silently fail to generate the correct code. (Encountered some additional cases, open an issue so that we can solve it or expand this list)

Imcomplete Layer Calls: All layer calls must be complete, i.e. the following is not allowed:

Lux.@layerdef function ThisModelDoesNotWork(x)
    @assert size(x, 1) == 2 && size(y, 1) == 2
    shared = Dense(2, 3, gelu)
    a = Dense(3, 3, relu)(shared(x) .+ Dense(2, 3, relu)(x)) .- shared(x)
    b = Dense(3, 2, sigmoid)(a)
    return b
end

Instead write it as:

Lux.@layerdef function ThisModelWorks(x)
    @assert size(x, 1) == 2 && size(y, 1) == 2
    shared = Dense(2, 3, gelu)(x)
    a = Dense(3, 3, relu)(shared .+ Dense(2, 3, relu)(x)) .- shared
    b = Dense(3, 2, sigmoid)(a)
    return b
end

Creating layers without =: All layer definitions must be preceeded by =. Anything else be it return, .+=. .=, etc. is not allowed.

# Invalid
Lux.@layerdef function WrongReturn(x)
    return Dense(2, 2)(x) .+ Dense(2, 2)(x)
end

Lux.@layerdef WrongOneLiner(x) = Dense(2, 2)(x) .+ Dense(2, 2)(x)

# Valid
Lux.@layerdef function CorrectReturn(x)
    x = Dense(2, 2)(x) .+ Dense(2, 2)(x)
    return x
end

Deviations from flax.linen.compact

This is heavily inspired by flax.linen.compact, though the way this works is entirely different from flax.

Unlike Flax, in Lux.@layerdef we still need to use the complete API to define the model. We don't perform automatic shape inference at this point. (This is a planned feature for future releases).

We don't allow specifying names for each layer.

Conditionals are not an issue

Shared Parameters / Layer Definition API

Use ← (type \leftarrow followed by a tab) to declare a new layer. For Example:

Lux.@layerdef function MySharedLayers((x, y))
    s1 ← Dense(2, 2)
    s2 ← Dense(2, 2)
    x = s1(x) .+ s2(x) .- s1(y)
    return x
end

Known Issues & Workarounds

If definining a @layerdef containing only 1 layer, there is a syntax error ERROR: syntax: function argument and static parameter names must be distinct around <line>. To workaround this, use @macroexpand and copy-paste the generated code

Warning

This is an experimental API. It is likely to have bugs and untested edge cases. If you find the generated code to be incorrect, please file an issue.

Showcase

SimpleRNN example rewitten using @layerdef

using Lux, Random, NNlib

Lux.@layerdef function SpiralClassifier(x::AbstractArray{T, 3}) where {T}
    lstm_cell ← LSTMCell(2 => 8)
    h, c = lstm_cell(view(x, :, 1, :))
    for i in 1:size(x, 2)
        h, c = lstm_cell(view(x, :, i, :), h, c)
    end
    y = Dense(8, 2, sigmoid)(h)
    return vec(y)
end

model = SpiralClassifier()

x = randn(Float32, 2, 4, 1)
ps, st = Lux.setup(Random.default_rng(), model)

model(x, ps, st)

(Float32[0.5464334, 0.6516966], (dense_1 = NamedTuple(), lstmcell_1 = (rng = Random.Xoshiro(0x08859d3b986ade14, 0xf3750638b584f6f1, 0x6941ec956ae609c5, 0x6c479b4c4706c2a0),)))

Index

Lux.@layerdef