Data Transfer

Lux.cpu — Function

cpu(x)

Transfer x to CPU

Lux.gpu — Function

gpu(x)

Transfer x to GPU

Initialization

Lux.glorot_normal — Function

glorot_normal(rng::AbstractRNG, size...; gain = 1)

Return an Array{Float32} of the given size containing random numbers drawn from a normal distribution with standard deviation gain * sqrt(2 / (fan_in + fan_out)). This method is described in [1] and also known as Xavier initialization.

References

[1] Glorot, Xavier, and Yoshua Bengio. "Understanding the difficulty of training deep feedforward neural networks." Proceedings of the thirteenth international conference on artificial intelligence and statistics. 2010.

Lux.glorot_uniform — Function

glorot_uniform(rng::AbstractRNG, size...; gain = 1)

Return an Array{Float32} of the given size containing random numbers drawn from a uniform distribution on the interval $[-x, x]$, where x = gain * sqrt(6 / (fan_in + fan_out)). This method is described in [1] and also known as Xavier initialization.

References

[1] Glorot, Xavier, and Yoshua Bengio. "Understanding the difficulty of training deep feedforward neural networks." Proceedings of the thirteenth international conference on artificial intelligence and statistics. 2010.

Lux.ones32 — Function

ones32(rng::AbstractRNG, size...) = ones(Float32, size...)

Return an Array{Float32} of ones of the given size. (rng is ignored)

Lux.zeros32 — Function

zeros32(rng::AbstractRNG, size...) = zeros(Float32, size...)

Return an Array{Float32} of zeros of the given size. (rng is ignored)

Index

Lux.cpu

Lux.glorot_normal

Lux.glorot_uniform

Lux.gpu

Lux.ones32

Lux.zeros32